+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1] - 2019-12-11
+### Added
+- DpiAwareness
+- Application ID column
+- support for DeploymentAction='OptionalInstallation'
+
+
 ## [1.0] - 2019-10-19
 ### Added
 - Added italian translation thx @irondave